night-orch: implement #57 DaisyUI and cleaner web ui
