--- web/postcss.config.cjs
@@ -1,10 +1,0 @@
-const path = require('node:path')
-
-module.exports = {
-  plugins: {
-    tailwindcss: {
-      config: path.join(__dirname, 'tailwind.config.cjs'),
-    },
-    autoprefixer: {},
-  },
-}
