+++ web/vite.config.ts
@@ -1,10 +1,11 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import tailwindcss from '@tailwindcss/vite'
 import { resolve } from 'node:path'
 
 export default defineConfig({
   root: resolve(import.meta.dirname),
-  plugins: [react()],
+  plugins: [tailwindcss(), react()],
   build: {
     outDir: 'dist',
     emptyOutDir: true,
